Feature: Add `BatchRename` script. Add .gitignore file
Add `BatchRename` script. A simple python utility script to randomly
rename a large number of files at once (without messing up the
extensions obviously).

I've made this script is in Visual Studio Code, and so I've adhered to the
Python documentation styling for VSCode. This means that the PyCharm guys
could be having some trouble with the documentation not being properly
visible for some parts.

Also, even I had to jump to PyCharm (VSCode is not as robust for
Python as I'd like), so I somehow ended up with a '.vscode' and '.idea'
directory for the same project -- as will be evident from the .gitignore
in this project.