BugFix: Include numbers in the random names if required
The boolean vairable was falling back to its default value which in turn
meant that numbers won't be included in the randomly generated strings
regardless of the user input.

Explicilty passing value for the last parameter does the trick and now
the files can have random numbers in their names if required.