Docs: Update the readme file with the latest changes
This commit updates the readme file with the changes being performed
in the previous commit (namely shifting to CLI mode).

Likewise, the previous documentation has been updated with the changes
being made, and how they affect the user. In particular, removed
parts that made the docs look confusing/complex.

Add example snippets on how to run this script with the various input
parameters.

Additionally, this commit also introduces minor change(s) in the functioning
of the script (specifically the backend functionality) - most of these being
refactoring code for optimization.